Fix user visual authentication bug when verifying email
When confirming an email address, the email verification app
passes user object to the template. This cause the base.html
template to call .is_authenticated on it and incorrectly conclude
that the user is logged in. I've fixed this by writing a custom
view described in the verification app readme:
https://github.com/LeoneBacciu/django-email-verification